+++ TIEcommerce/Sources/Filters/FiltersViews/RangeFilters/Models/Appearance/RangeFilterLayout.swift
@@ -1,0 +1,4 @@
+public enum RangeFilterLayout {
+    case textFieldsOnTop
+    case textFieldsFromBelow
+}