+++ TIEcommerce/Sources/Filters/FiltersViews/RangeFilters/Models/Formatter/RangeValuesFormatterProtocol.swift
@@ -1,0 +1,4 @@
+public protocol RangeValuesFormatterProtocol {
+    func string(fromDouble value: Double) -> String
+    func double(fromString value: String) -> Double
+}