+++ TIEcommerce/Sources/Filters/FiltersViews/RangeFilters/Protocols/RangeFilterDelegate.swift
@@ -1,0 +1,4 @@
+public protocol RangeFilterDelegate: AnyObject {
+    func valuesIsChanging(_ value: FilterRangeValue)
+    func valueDidEndChanging(_ value: FilterRangeValue)
+}