+++ TIEcommerce/Sources/Filters/FiltersViews/RangeFilters/Views/FilterRangeView/FilterRangeViewRepresenter.swift
@@ -1,0 +1,4 @@
+public protocol FilterRangeViewRepresenter: AnyObject {
+    func configureTextFields(with value: FilterRangeValue)
+    func configureRangeView(with value: FilterRangeValue)
+}