+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **Add**: Tag like filter collection view
 - **ADD**: List like filter table view
+- **ADD**: Range like filter view
 
 ### 1.26.0